Extraigo la clase CustomInfoProvider a su propio archivo
Extraigo la clase CustomInfoProvider para que pueda
ser usada por varias test suites

    refs:#18

    refs:#18
